Fixing bug #1
git-svn-id: https://apertium.svn.sourceforge.net/svnroot/apertium@668 72bbbca6-d526-0410-a7d9-f06f51895060

+++ lttoolbox/lttoolbox/FSTProcessor.C
@@ -627,7 +627,8 @@ FSTProcessor::analysis(FILE *input, FILE *output)
 
         unsigned int limit = sf.find(' ');
         unsigned int size = sf.size();
-        limit = (limit == string::npos?size:limit);
+        limit = (limit == static_cast<unsigned 
+int>(string::npos)?size:limit);
         input_buffer.back(1+(size-limit));
         printUnknownWord(sf.substr(0, limit), output);
       }
@@ -635,7 +636,8 @@ FSTProcessor::analysis(FILE *input, FILE *output)
       {
         unsigned int limit = sf.find(' ');
         unsigned int size = sf.size();
-        limit = (limit == string::npos?size:limit);
+        limit = (limit == static_cast<unsigned 
+int>(string::npos)?size:limit);
         input_buffer.back(1+(size-limit));
         printUnknownWord(sf.substr(0, limit), output);
       }
@@ -1512,7 +1514,8 @@ FSTProcessor::SAO(FILE *input, FILE *output)
 
         unsigned int limit = sf.find(' ');
         unsigned int size = sf.size();
-        limit = (limit == string::npos?size:limit);
+        limit = (limit == static_cast<unsigned 
+int>(string::npos)?size:limit);
         input_buffer.back(1+(size-limit));
         fwrite_unlocked("<d>", 3, sizeof(char), output);
         fwrite_unlocked(sf.c_str(), limit, sizeof(char), output);
@@ -1523,7 +1526,8 @@ FSTProcessor::SAO(FILE *input, FILE *output)
 
         unsigned int limit = sf.find(' ');
         unsigned int size = sf.size();
-        limit = (limit == string::npos?size:limit);
+        limit = (limit == static_cast<unsigned 
+int>(string::npos)?size:limit);
         input_buffer.back(1+(size-limit));
         fwrite_unlocked("<d>", 3, sizeof(char), output);
         fwrite_unlocked(sf.c_str(), limit, sizeof(char), output);
